+++ service/index.js
@@ -2,8 +2,10 @@
 
 const LeagueService = require('./league/league-service');
 const ManagerService = require('./manager/manager-service');
+const TeamService = require('./team/team-service');
 
 module.exports = {
     LeagueService,
     ManagerService,
+    TeamService,
 };