+++ service/manager/manager-service.js
@@ -5,6 +5,7 @@ const Logable = require('../../lib/log/logable');
 const { ManagerModel } = require('../../model/index');
 
 /** @typedef {import('../../lib/controller/controller').LoggingContext} LoggingContext */
+
 /**
  * @typedef {Object} ManagerData
  * @property {number} userId -
@@ -12,7 +13,6 @@ const { ManagerModel } = require('../../model/index');
  * @property {string} lastName -
  * @property {string} username -
  * @property {string} [secret] -
- * @property {string} leagueId -
  */
 
 class ManagerService extends Logable {
@@ -109,35 +109,32 @@ class ManagerService extends Logable {
         await this._managerModel.patch(ctx, { user });
     }
 
-    getLeagueId(ctx, options) {
-        const {
-            manager,
-        } = options;
+    async joinLeague(ctx, { userId, leagueId }) {
+        this.logDebug(ctx, 'Joining a manager to a league.', { userId, leagueId });
 
-        return manager.leagueId;
+        try {
+            await this._managerModel.patch(ctx, { user: { userId, leagueId } });
+        } catch (error) {
+            this.logError(ctx, 'Error on joining a manager to a league.', { error });
+
+            throw error;
+        }
+
+        this.logDebug(ctx, 'A manager is joined to a league.', { userId, leagueId });
     }
 
-    hasLeague(ctx, options) {
-        const {
-            manager,
-        } = options;
+    async leaveLeague(ctx, { userId }) {
+        this.logDebug(ctx, 'Unjoining a manager from a league.', { userId });
 
-        return !!manager.leagueId;
-    }
+        try {
+            await this._managerModel.patch(ctx, { user: { userId, leagueId: null } });
+        } catch (error) {
+            this.logError(ctx, 'Error on unjoining a manager from a league.', { error });
 
-    async joinLeague(ctx, options) {
-        const {
-            userId,
-            leagueId
-        } = options;
+            throw error;
+        }
 
-        await this._managerModel.patch(ctx, { user: { userId, leagueId, }});
-    }
-
-    async leaveLeague(ctx, options) {
-        const { userId } = options;
-
-        await this._managerModel.patch(ctx, { user: { userId, leagueId: null } });
+        this.logDebug(ctx, 'A manager has left a league.', { userId });
     }
 }