+++ transport/bot/controller/command/join-league/join-league-command-controller.js
@@ -8,16 +8,13 @@ class JoinCommandController extends Controller {
         super(options);
 
         this._leagueService = options.leagueService;
-        this._managerService = options.managerService;
     }
 
     async process(ctx, { request }) {
         const {
             state: {
                 manager,
-            },
-            message: {
-                text,
+                league,
             },
         } = request;
 
@@ -25,40 +22,48 @@ class JoinCommandController extends Controller {
             userId,
         } = manager;
 
-        const secret = this._getText(ctx, { request });
+        if (league) {
+            const { name } = league;
 
-        if (!secret) {
-            this.logWarn(ctx, 'Unable to join a user to a league without a secret.');
-            request.reply('Please, sent a league secret. Use /join-league [secret] command.');
+            request.reply(`You are already in ${name} league. To leave a league use /leaveleague command.`);
 
             return;
         }
 
-        if (this._managerService.hasLeague(ctx, { manager })) {
-            request.reply('You have already in a league.');
-            request.reply('Leave a league with /leave-league command first.');
+        this.log(ctx, 'Joining a user to a league.', { userId });
+
+        const secret = this._getText(ctx, { request });
+
+        if (!secret) {
+            this.logWarn(ctx, 'Unable to join a user to a league without a secret.', { userId });
+            request.reply('Please, sent a league secret. Use /joinleague [secret] command.');
 
             return;
         }
 
         this.log(ctx, 'Joining a user to a league.', { userId, secret })
 
-        const league = await this._leagueService.getBySecret(ctx, { secret });
+        const userLeague = await this._leagueService.getBySecret(ctx, { secret });
 
-        if (!league) {
-            request.reply('A league with secret ' + secret + ' is not found.');
-            request.reply('Create a league with /create-league [name] command or use another secret.');
+        if (!userLeague) {
+            this.logWarn(ctx, 'Unable to join a user to a league with invalid secret.', { userId });
+
+            request.reply(`A league with secret ${secret} is not found.`);
+            request.reply('Create a league with /createleague [name] command or use another secret.');
             
             return;
         }
 
-        const { leagueId, name } = league;
+        const { leagueId, name } = userLeague;
 
-        await this._managerService.joinLeague(ctx, { userId, leagueId });
+        await this._leagueService.joinManager(ctx, {
+            userId,
+            leagueId,
+        });
 
         request.reply(`You have joined a league [${name}].`);
 
-        this.log(ctx, 'A user joined to a league.', { userId, leagueId });
+        this.log(ctx, 'A user is joined to a league.', { userId, leagueId });
     }
 }